+++ talpid-core/src/dns/linux/systemd_resolved.rs
@@ -54,7 +54,13 @@ pub enum Error {
     MatchDBusTypeError(#[error(cause)] dbus::arg::TypeMismatchError),
 }
 
-const DYNAMIC_RESOLV_CONF_PATH: &str = "/run/systemd/resolve/resolv.conf";
+lazy_static! {
+    static ref RESOLVED_PATHS: Vec<&'static Path> = vec![
+        Path::new("/run/systemd/resolve/stub-resolv.conf"),
+        Path::new("/run/systemd/resolve/resolv.conf"),
+    ];
+}
+
 const RESOLVED_DNS_SERVER_ADDRESS: [u8; 4] = [127, 0, 0, 53];
 
 const RESOLVED_BUS: &str = "org.freedesktop.resolve1";
@@ -103,12 +109,31 @@ impl SystemdResolved {
 
     fn ensure_resolv_conf_is_resolved_symlink() -> Result<()> {
         let is_correct_symlink = fs::read_link(RESOLV_CONF_PATH)
-            .map(|resolv_conf_target| resolv_conf_target == Path::new(DYNAMIC_RESOLV_CONF_PATH))
+            .map(|resolv_conf_target| Self::compare_resolvconf_symlink(&resolv_conf_target))
             .unwrap_or_else(|_| false);
         if is_correct_symlink {
             Ok(())
         } else {
             Err(Error::NotSymlinkedToResolvConf)
+        }
+    }
+
+    fn compare_resolvconf_symlink(link_path: &Path) -> bool {
+        // if link path is relative to /etc/resolv.conf, resolve the path and compare it.
+        if link_path.is_relative() {
+            match Path::new("/etc/").join(link_path).canonicalize() {
+                Ok(link_destination) => RESOLVED_PATHS.contains(&link_destination.as_ref()),
+                Err(e) => {
+                    log::error!(
+                        "Failed to canonicalize resolv conf path {} - {}",
+                        link_path.display(),
+                        e
+                    );
+                    false
+                }
+            }
+        } else {
+            RESOLVED_PATHS.contains(&link_path)
         }
     }