+++ CHANGELOG.md
@@ -60,6 +60,9 @@ Line wrap the file at 100 chars.                                              Th
 #### Windows
 - Add better offline detection
 
+#### Linux
+- Fix `systemd-resolved` detection by better compraing `/etc/resolv.conf` symlinks.
+
 
 ## [2019.4] - 2019-05-08
 This release is identical to 2019.4-beta1